+++ expediagroup-sdk-core/src/main/kotlin/com/expediagroup/sdk/core/auth/common/Credentials.kt
@@ -1,0 +1,3 @@
+package com.expediagroup.sdk.core.auth.common
+
+interface Credentials