+++ expediagroup-sdk-core/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [0.0.6-alpha](https://github.com/ExpediaGroup/expediagroup-java-sdk/compare/expediagroup-sdk-core-v0.0.5-alpha...expediagroup-sdk-core-v0.0.6-alpha) (2025-03-26)
+
+
+### Features
+
+* mask sensitive auth fields ([#1086](https://github.com/ExpediaGroup/expediagroup-java-sdk/issues/1086)) ([89a9f73](https://github.com/ExpediaGroup/expediagroup-java-sdk/commit/89a9f733e42466119812c59326d2aef3cc10cfb0))
+
+## [0.0.5-alpha](https://github.com/ExpediaGroup/expediagroup-java-sdk/compare/expediagroup-sdk-core-v0.0.4-alpha...expediagroup-sdk-core-v0.0.5-alpha) (2025-03-26)
+
+
+### Features
+
+* add Credentials marker interface ([#1084](https://github.com/ExpediaGroup/expediagroup-java-sdk/issues/1084)) ([65fa5f9](https://github.com/ExpediaGroup/expediagroup-java-sdk/commit/65fa5f91182f0b4ff66994e622a2f1b60e3f7978))
+
+## [0.0.4-alpha](https://github.com/ExpediaGroup/expediagroup-java-sdk/compare/expediagroup-sdk-core-v0.0.3-alpha...expediagroup-sdk-core-v0.0.4-alpha) (2025-03-25)
+
+
+### Features
+
+* support Basic and OAuth credentials types ([#1074](https://github.com/ExpediaGroup/expediagroup-java-sdk/issues/1074)) ([0970108](https://github.com/ExpediaGroup/expediagroup-java-sdk/commit/097010870fb4fb17c53f9d5b87d30f0b43e3658a))
+
 ## [0.0.3-alpha](https://github.com/ExpediaGroup/expediagroup-java-sdk/compare/expediagroup-sdk-core-v0.0.2-alpha...expediagroup-sdk-core-v0.0.3-alpha) (2025-03-09)